TST: Reconcile (python) tests with #172 changes
honestly, this was really annoying to do, but at least it's done.

Also removed the one print message (warning about dropped samples),
and replaced it with a logging.warning() call for consistency's sake.
This will make it harder to test this, so I just (temporarily?)
removed the part of the integration test that looks for this message
to be present (it's clear that the logic *works*).