BUG: ensure that empty f. metadata vals are None
This was already being converted to null by altair, but Pandas was
setting it to NaN (yikes).

Of course, since Altair converts NaNs to Nones, this shouldn't have
affected anything. But IMO it's good to make sure we're consistently
sending Nones to Altair (which we *know* will be converted to null
in JSONs).