BUG/MAINT: Fix+simplify setting default balances
Now, we avoid using pd.merge() and instead just directly create a
qurro_balance column in sample_metadata (checking first, of course,
that no such column already exists).

We also take care to initialize the qurro_balance column as None,
instead of NaN as was being done before. Not that this really matters
too much -- Altair converts NaNs to Nones anyway -- but good to
be consistent.

Also added on type annotations in the sample plot Altair definition
to ensure that Altair still treats qurro_balance as a numeric
field (even though it might infer it as a nominal field due to the
Nones, or due to the object dtype that pandas applies to it due to
said Nones).