ENH: Remove empty features and add tests (#171)
This has apparently thrown off the Byrd test, since it looks like
there were a bunch of empty features in that dataset. Need to update
the testing utilities to allow for empty features to be removed.

also looks like the JSONs for the q2-moving-pictures and sleep apnea
integration tests are different -- figure out why, and if that's due
to a bug or just due to something else. once that's done this issue will
be done