MAINT/BUG: Simplify+clear up support_files copying
The approach here was based on Emperor's. Turns out shutil.copytree()
is inherently limited (the destination directory is not allowed to
exist at all), which is a problem for the Q2 plugin for Qurro.
(See https://stackoverflow.com/q/1868714/10730311 for lots of
discussion on this problem -- the use of distutils.dir_util.copy_tree
is actually brought up here as well.)

Anyway, this works! Code is much simpler than before, so this
closes #69. (Note that the new approach doesn't bother filtering out
.DS_STORE stuff, but this should be ok since those sorts of garbage
files shouldn't be inclued in Qurro at all since they aren't specified
in the MANIFEST.in file. And even if they are they won't do any harm.)

Also, this makes Qurro's file overwriting consistent (#79). Qurro's Q2
plugin behaves how Q2 wants it to (QIIME 2 manages all of the actual
writing stuff of the .qzv files), so we can do whatever works best
for the standalone Qurro tool. And the use of copy_tree() now means
that outputs are overwritten: if you have an existing directory named
"x" and specify it as the output dir of standalone Qurro, Qurro will
write the visualization contents to inside x, no sweat.

If "x" contains other stuff besides Qurro visualizations, then
copy_tree() won't overwrite that stuff. If "x" contains a former Qurro
visualiztion (or just stuff with conflicting file names), then that
stuff will be naturally overwritten.

This is all reasonable behavior, so #79 is mostly taken care of. The
remaining TODOs for that are 1) documenting this somewhere for
standalone Qurro and 2) maybe adding some tests for this (but this is
all just normal behavior of copy_tree() so we don't really need to
worry about thoroughly testing it).