BUG: Only bail out if input text is actually empty
what was happening was that you could enter in something like " " as
the input text, and it'd just get rejected (because the previous code
immediately bailed out of searching if inputText.trim().length was 0).

However, text searching uses the *exact* text you enter in, whitespace
and all. This behavior is ok, but it means that we shouldn't
necessarily filter out inputs containing only whitespace (since the
user could conceivably want to find all features containing a space
for a given metadata field, and exact text searching should provide
that functionality).

We still check to make sure that the input text isn't empty, so #3 is
still resolved. And the rank searching trims the input text anyway, so
all this change affects is the text searching controls.