MAINT: add dom_utils.getNumberIfValid() and use
This abstracts some repeated work that came up in the process of
implementing #141.

I might make this be used by getInvalidSampleIDs() in the future.
There are some arguments for and against doing that (in short:
for = maximum consistency, against = possibility for inconsistency
with the vega-lite filtering expressions)... will think about it