ENH: --verbose -> --debug; make available in Q2
This closes #95 (not the way I expected to handle this, but this is
IMO a solid solution).

Basically, you can say --verbose in Q2 to see the messages that would
be output normally when running Qurro standalone. And in either the
Q2 or non-Q2 Qurro versions, you can use --(p-)debug to see debug
messages as well.