DOC: Add mackerel demo + link to analysis!
This is needed for the Qurro paper.

Once Qurro v0.3.0 is merged into the biocore repo, I can merge these
changes with my changes (and also do cool stuff like add the mackerel
data to the Qurro test suite, so its demo is continuously updated
as new versions of Qurro are released).