MNT: Change up viz formatting; #189 interface
I made the buttons look a bit nicer (IMO), moved the export sample
plot data button to be with the sample plot controls, changed
"Regenerate plots" to "Apply Filtering" to be clearer, and overall
started changing up the viz interface in anticipation of implementing
the functionality described in #189.

Oh, also! I removed the #marginOnTop CSS class, since it turns out
that Bootstrap already provides that via the mt-* classes. NICE