BUG/TST: updateFeaturesText... b4 updateLogRatio
Since the updateLogRatio() call is awaited, but the
updateFeaturesText...() call just sort of happens afterwards, this
was causing minor problems with tests awaiting plot regeneration *but*
the DOM not having been updated yet by the time the test is executed.
(That's my understanding of the situation at least.)

this should fix that, albeit at the "cost" of
the features text displays getting updated slightly before the plots
are updated (doesn't really matter IMO, going to be one-then-the-other
either way).

A neat side effect of this is that this lets us repeat this one
multi-feature thing before *each* test instead of just "before" a
group of tests. hunting down this bug sucked but at least it's done
now :)

not gonna bother updating the changelog since this doesn't change
anything from a user perspective, i think