MNT/BUG: make autoselecting logistics more robust
TLDR: in the previous code, it was possible to pass in a garbage
search type (e.g. "autoASDF"), which would end up in the
auto-selection block in filterFeatures but would mistakenly get
assumed to be autoLiteral searching.

This explicitly checks to make sure the search type is one of four
valid types, so other types will get an error by default.

also changed a few small things up (pre-computing "inPercentages"
in order to only have to do the .startsWith() call once). Code should
be more legible now.