REL: Update mackerel test data re latest analysis
This is a duplicate of commit 7765fd4950ebe1adff0d82097c8345b50b9a534a.
I'm doing it directly here so we can apply these changes ASAP, since
I'm going to be presenting on this data soon and I don't want this
to be blocked by having to finish up all of the other stuff in
my and Antonio's forks.