ENH/BUG: basic numeric range checking #242
This does not fully address #242, since it's still possible to pass
in arbitrarily large numbers through sample/feature metadata, and this
could theoretically break things in the visualization interface.

As the docs for vibe_check() note, checking for those cases (given
the ridiculous amount of variety in metadata) is going to be a pain.

However, this solution should at least address most of the scary cases
where this problem could result in incorrect results. It also
ties in nicely with #250!