MNT: Clean up some of the DataTables code; tst fix
This doesn't completely resolve the errors we get when running
the JS test from DataTables, but it at least seems to knock out one
of the types of errors (something yelling about not being able
to reinitialize a DT). Calling .destroy() in the RRVD destroy()
function seems to bypass this problem. i think.