Fixed bug where no common samples test was failing
Update to biom caused some issue with the interaction of pandas and
biom. Filtering the taxonomy joined df to the numerator/denominator
features indices would result in any columns of only 0s to be converted
to NaNs. Because these columns will never contribute to the log-ratios,
we can just discard them in the beginning.