DOC: add qurro usage stuff back to color example
A lot of this was based on @cameronmartino's excellent initial version
of this tutorial -- many thanks.

This includes some screenshots, analogous to the TCGA tutorial.
These screenshots will already be slightly "out of date" in terms
of the UI when newer versions of Qurro come out, but honestly I
don't really think that's a big enough deal to justify time
integrating emperor mchelper into this -- kinda swamped rn