MNT: new utility func for creating test RRVDisplay
the new function avoids the JSON-overwriting bug mentioned in the
prev commit, and is pretty easy to use. I went ahead and replaced
a bunch of the tests' code that created new RRVDisplays manually
with calls to this utility function.

a pleasant side effect of that change is that i was able to remove
a requirement of display.js to be loaded in many of the tests --
saves a tiny bit of space? maybe? not really