ENH: Add sample border checkbox #240
This adds basically all of the functionality for #240, but there's
still a lot of work to be done on 1) handling corner cases
(interactions with the boxplot will be annoying, and should be well-
tested) and 2) adding tests for this stuff.

But yeah this is already pretty nice! It's super useful for faint/
pastel color schemes