TST: test polyphyly example for #224; imprv errmsg
The error message improved has to do with what's thrown when
a nonexistent field name is passed to filterFeatures(). Should never
be seen by the user, but having a more detailed error message (that
includes the name of said nonexistent field) will help with debugging.
(I also modified the test for that error message to check that
the error message is actually thrown.)