DOC: Move selection appendix to sep doc from MPtut
To give a unified space to document the whitespace junk, as well
as other quirks about these methods. Basically does what I was trying
to get at with #123.

Would be good to modify tooltips in interface to just point to this
explanation of these methods in lieu of trying to explain the details
of the sep. fragment searching there...