MNT/BUG/ENH: scale-w-page "constant plot width"
see comments. basically, rather than getting the .width from the
rank plot JSON, we now compute a reasonable upper bound width
on the rank plot from the page, and then scale the bars accordingly.

it's not gonna look the SAME as before but i think it is actually
a nicer solution :)