ENH: add jitter checkbox to sample plot (fix #139)
Took less than six years so i think this is officially Fast (TM)

Need to do more testing to verify this looks good. Mainly,
I want to remove the xOffset encoding stuff from the JSON
specification entirely when the x-axis is set to a quantitative
scale. It doesn't LOOK like the xoffset impacts quantitative plots,
thankfully, but I don't trust it to not break later on. I really
don't want this to subtly introduce noise into quantitative plots
where the x-axis position actually means something

but yeah this is great! i think the next big feature that would
be nice is making the sample plot bigger, because my old tired eyes
are like yeah these are tiny points. but that is lower priority