Unpin Altair and SciPy; update Vega* versions; support jitter plots (#331)
* DOC: yeah it's been like five years

this is no longer a "temporary" caveat. also "caveat" was a bad
word to use to describe this, we're not like running a gladitorial
coliseum here or something.

* DOC: PyPI / conda-forge badge colors

not all that important, but while we're fixing things...

* ENH: make "qurro" and "qurro -h" work nicely

* DEP: unpin altair, scipy, python: close #216

Looks like there are breaking API changes to altair, so gotta fix
those next (and do a postmortem of "okay how badly does this break
OLD versions of altair" / "is it worth trying to support those or do
we just tell users to install an old version of qurro if needed")

* DEP/MNT: get Python side working with new versions

For #216. Need to at minimum update JS side to adjust bar
width slider

* STY: minor updates to py code re: new flake8/black

* MNT/BUG: get bar width slider working in JS

Need to update the checkbox. Bizarrely, it seems like
rankPlotJSON.config.view.width does not exist any more -- instead,
there is "continuousWidth", which is smaller for some reason ????

We can still SET a width / height directly, though, even though
altair does not let you do that (but modifying the JSON dict after
the fact is apparently cool with vega).

Ideally i minimize the amount of dreadful jank workarounds i have
to make here, but also i just ... want the old width back. why is
this so complicated

* ENH: use labelSeparation in rank plot x-axis :)

* MNT/BUG/ENH: scale-w-page "constant plot width"

see comments. basically, rather than getting the .width from the
rank plot JSON, we now compute a reasonable upper bound width
on the rank plot from the page, and then scale the bars accordingly.

it's not gonna look the SAME as before but i think it is actually
a nicer solution :)

* TST: partially update tests

* BUG/MNT: use autocomplete="off" on checkboxes

* DOC/MNT: upd8s re: divine punishment; labelsep+

for the mackerel example labelseparation >= 3 looks nicer than 2

* STY: index html

* MNT/BUG: escape . -> _ and : -> ; now #66

since altair now reserves :

* ENH: use paddingOuter when scaling rank plot

only really helps w matching test, i think, but still nice

* TST: update tests re changes

note the changes to test_rrvdisplay.js wrt the new fixed size stuff

* STY: prettify recently-changed JS

* TST: more JS tests re new padding stuff

* TST: more fixed width/etc tests

* DEP/TST/BLD: supporting py 3.6+ now???

let's see if this works ....

* STY: lingering js test style fixes

* BLD/TST: try building w/ all Q2 envs >= 2020.11

i am not sure this will work, but... let's ese

* TST: put python versions in quotes

to prevent 3.10 from being read as 3.1 apparently. forgot
that was something we had to do. god is dead

* TST: try ubuntu 20.04 for q2 builds

since they use py 3.6, that might be the problem?

* BLD: try trimming q2 CI for now to 2022.11+

since CI straight up can't install old versions of QIIME 2 now

* BLD: try increasing miniconda action thing vsn?

since old q2 envs are now literally uninstallable apparently

if this doesnt work just trim the old ones for now

* STY/TST: update demos

* DOC: update fit bar width tooltip

* ENH: add jitter checkbox to sample plot (fix #139)

Took less than six years so i think this is officially Fast (TM)

Need to do more testing to verify this looks good. Mainly,
I want to remove the xOffset encoding stuff from the JSON
specification entirely when the x-axis is set to a quantitative
scale. It doesn't LOOK like the xoffset impacts quantitative plots,
thankfully, but I don't trust it to not break later on. I really
don't want this to subtly introduce noise into quantitative plots
where the x-axis position actually means something

but yeah this is great! i think the next big feature that would
be nice is making the sample plot bigger, because my old tired eyes
are like yeah these are tiny points. but that is lower priority

* PERF: remove old HTML comment

* REL: draft changelog

* DEP: force newer skbio vsn for standalone CI

* DEP: just trying things

wats going on

* DEP/BLD: limit standalone testing to py >= 3.8

* BUG/DEP: exorcise np.NaN in favor of np.nan

now THIS is programming

* BLD: attempt to fix node.js installation in build

okay we're making progress!!!

node 12 seems really out of date but its wat i have on my laptop.
eventually do binary search or whatever to figure out what if
anything broke

* MNT: replace distutils with shutil for new py vsns

* BLD: try removing some old q2 vsns & pinning nyc

see if this fixes builds

ideally ofc we'd add back automated testing on all q2 versions
but if needed we can limit focus of recent release to new q2 vsns

* DEP: try bumping altair pin to >= 4

* BLD: fail fast = false in standalone CI

* REL: changelog updates

* BLD: let's see how all (ish) supported q2 vsns do

* BLD: update q2 CI Node.js stuff

this should really be standardized between the two CI files eventually

* BLD: skbio vsn workaround for py 3.6,3.7

* DEP: use distutils' copy_tree() for Python < 3.8

out with the old, in with the new, except for not really

* BLD: remove the {{ }} in standalone CI if stmt

does this appease the yaml gods? maybe.

* BLD: use single-quotes, apparently???

* TST: accept old OR new Vega-Lite mark bar stuff

* STY

* BLD: pin skbio = 0.5.5 for py 3.6/3.7?

* BLD: more numpy/scikitbio jank for old py versions

* BLD: separate cyth/np installs from skbio one

already was doing but not in prev commit's workaround

* BLD: for now, rm q2 2022.2 & 2022.8 from ci matrix

see comments. no idea why these versions in particular are janking out

* BLD: try bumping checkout & setup-python vsns

maybe this fixes the python 3.7 installing numpy issue?

* BLD: try pinning py 3.7.7 specifically???

* BLD: nvm pinning python to 3.7.7 didn't work

* BLD: install numpy for py < 3.8 w/ apt?

i'll try anything at this point....

* BLD: try -y?

* BLD: just throwing spaghetti at the wall

* DOC: tooltip wording tweaks

* BLD: try pinning skbio 0.5.5 but not numpy??

* BLD: okay now pin numpy top vsn for py < 3.8 blds

i am becoming the joker

* BLD: try installing py36/py37 using conda

Insanely jank solution but if it works then great

* BLD: whoops remove old if stmt

* BLD: ohhhh wrap pip installs in "conda run"s

will this work??? maybe???

* BLD: conda jank

* DOC/BLD: Python 3.7 is working now!!! yay

what a jank solution. i am sure it will break sometime, but for
now we get to keep officially supporting python 3.6/3.7 :)

* MNT: Adjust df assignments to appease pandas #332

* MNT: more .loc fixes for #332

* MNT: .dtypes[col idx] --> .dtypes[col id] #332

* MNT/TST: update replace_nan() & tests re: None/NaN

Closes #332.

The logic here is kind of finicky, but I think this should have
the desired effect -- of missing values in the input sample/feature
metadata ending up consistently as nulls in the JSON.

* MNT/TST: whoops fix another warning #332

* MNT/BUG: better jitter updates #135

- add jitter on quant -> cat scale chg if needed
- remove jitter on cat -> quant scale chg if needed
- uncheck jitter checkbox in destroy()

* TST: jitter checkbox enabling/disabling w/ boxplot

* TST: jitter stuff updates JSON properly

* DOC: update demos

* REL/DOC: setup vsn py classifiers; Q2 vsns